Fixes according to comments
I've changed the code according to your comments. The only thing that seems to be broken is the Paper wrap on the Favicon.

I still need to add Favicon story to render it in storybook, that my goal for tomorrow. Also I'm kinda confused what test to write for the SourceItem component (maybe maxChar test?)