#48_source_component (#50)
* #48_source_component

Closes #48

Problems:
- I can't figue it out how to add margin/padding so it will looke like the one in Figma
- Date font doesn't look right even thou it's Roboto as in Figma
- Icon is rly bad looking with current size and it's not central placed (it kinda jumps all over and I cannot seem to grasp controll over it)
- SupportIndicator will be changed in the next commit

* #48_source_item

Fix on Support Indicator

* Fixes according to comments

I've changed the code according to your comments. The only thing that seems to be broken is the Paper wrap on the Favicon.

I still need to add Favicon story to render it in storybook, that my goal for tomorrow. Also I'm kinda confused what test to write for the SourceItem component (maybe maxChar test?)

* requested fixes + favicon story

Now with Favicon story with two variants should we also have two variants of the SourceItem component?

* updated favicon varian

That should close #48 and #36

* Add some small fixes

- layout: centered only for SourceItem story, this was messing with other components.
- removed circular favicon wrapper
- removed some commented code

Co-authored-by: Tobias Tykvart <tobias.tykvart@posteo.de>
Co-authored-by: Tobias Tykvart <tobias@factiverse.no>